[Smoke Tests] Unify test_utils.rs and utils.rs
