+++ testsuite/smoke-test/src/aptos_cli/validator.rs
@@ -3,7 +3,7 @@
 
 use crate::{
     smoke_test_environment::SwarmBuilder,
-    test_utils::{create_and_fund_account, MAX_CATCH_UP_WAIT_SECS},
+    utils::{create_and_fund_account, MAX_CATCH_UP_WAIT_SECS},
 };
 use aptos::{
     account::create::DEFAULT_FUNDED_COINS,
