+++ testsuite/smoke-test/src/full_nodes.rs
@@ -4,7 +4,7 @@
 
 use crate::{
     smoke_test_environment::SwarmBuilder,
-    test_utils::{
+    utils::{
         assert_balance, create_and_fund_account, transfer_coins, MAX_CATCH_UP_WAIT_SECS,
         MAX_CONNECTIVITY_WAIT_SECS, MAX_HEALTHY_WAIT_SECS,
     },
