+++ testsuite/smoke-test/src/txn_broadcast.rs
@@ -3,7 +3,7 @@
 
 use crate::{
     smoke_test_environment::SwarmBuilder,
-    test_utils::{
+    utils::{
         assert_balance, create_and_fund_account, transfer_coins, MAX_CATCH_UP_WAIT_SECS,
         MAX_CONNECTIVITY_WAIT_SECS, MAX_HEALTHY_WAIT_SECS,
     },
